pyproject.toml
[tool.poetry]
name = "pureapi"
version = "3.2.0"
description = "Tools for working with Elsevier's Pure API."
readme = "README.md"
authors = ["David Naughton <naughton@umn.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.3"
addict = "*"
requests = "^2.20.0"
tenacity = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-forked = "*"
python_dotenv = "*"
